+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.0.1 - 2017-12-29
+### Added
+* Added cache busting of the Route Map caches when Elements are saved
+* Added Route Map Cache as a cache that can be cleared via the Clear Caches utility
+
 ## 1.0.0 - 2017-12-27
 ### Added
-- Initial release
+* Initial release